+++ Assets/DeltaDNA/Editor/Android/Menus/MenuItems.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright (c) 2016 deltaDNA Ltd. All rights reserved.
+// Copyright (c) 2017 deltaDNA Ltd. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
 // limitations under the License.
 //
 
+using System.IO;
 using UnityEditor;
 using UnityEngine;
 
@@ -52,6 +53,11 @@ namespace DeltaDNA.Editor {
                     true)
                     .Show();
             }
+        }
+
+        internal static bool AreAndroidNotificationsInProject() {
+            return File.Exists("Assets/Plugins/Android/deltadna-sdk-unity-notifications")
+                && File.Exists("Assets/DeltaDNA/Plugins/Android");
         }
     }
 }